pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stc-framework"
version = "0.2.0"
description = "Stalwart-Trainer-Critic: production-grade AI agent system framework with zero-trust governance, optimization, and data sovereignty."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "STC Framework Contributors" }]
keywords = ["ai", "agents", "governance", "llm", "rag", "guardrails", "aiuc-1"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "structlog>=24.1",
    "tenacity>=8.2",
    "httpx[http2]>=0.26",
    "opentelemetry-api>=1.22",
    "opentelemetry-sdk>=1.22",
    "prometheus-client>=0.19",
    "numpy>=1.24",
    "cryptography>=42.0",
]

[project.optional-dependencies]
presidio = ["presidio-analyzer>=2.2", "presidio-anonymizer>=2.2", "spacy>=3.7"]
qdrant = ["qdrant-client>=1.7"]
litellm = ["litellm>=1.50"]
langfuse = ["langfuse>=2.0"]
phoenix = ["arize-phoenix>=5.0"]
nemo = ["nemoguardrails>=0.10"]
guardrails-ai = ["guardrails-ai>=0.5"]
lightning = ["agentlightning"]
parquet = ["pyarrow>=14.0"]
langgraph = ["langgraph>=0.2", "langchain-core>=0.3"]
otlp = ["opentelemetry-exporter-otlp>=1.22"]
service = ["flask>=3.0", "gunicorn>=21.2", "flask-limiter>=3.5", "werkzeug>=3.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "hypothesis>=6.92",
    "ruff>=0.3",
    "black>=24.1",
    "mypy>=1.8",
    "respx>=0.20",
    "freezegun>=1.4",
    "types-PyYAML",
]
all = [
    "stc-framework[presidio,qdrant,litellm,langfuse,phoenix,nemo,guardrails-ai,parquet,langgraph,otlp,service]",
]

[project.urls]
Homepage = "https://github.com/stc-framework/stc-framework"
Documentation = "https://github.com/stc-framework/stc-framework/tree/main/docs"
Issues = "https://github.com/stc-framework/stc-framework/issues"

[project.scripts]
stc-agent = "stc_framework.reference_impl.scripts.run_agent:main"
stc-baseline = "stc_framework.reference_impl.scripts.run_baseline:main"
stc-red-team = "stc_framework.adversarial.runner:main"
stc-governance = "stc_framework.governance.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["stc_framework*"]

[tool.setuptools.package-data]
stc_framework = ["spec/default_spec.yaml", "py.typed"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: integration tests requiring external services (opt-in)",
    "slow: tests that take more than a second",
]

[tool.coverage.run]
source = ["src/stc_framework"]
branch = true
omit = [
    "*/reference_impl/scripts/*",
    "*/adversarial/runner.py",
    "*/service/wsgi.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "\\.\\.\\.",
]
precision = 1
show_missing = true

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF", "N", "D"]
ignore = ["D100", "D101", "D102", "D103", "D104", "D105", "D107", "D203", "D213"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "qdrant_client.*",
    "presidio_analyzer.*",
    "presidio_anonymizer.*",
    "litellm.*",
    "langfuse.*",
    "phoenix.*",
    "nemoguardrails.*",
    "guardrails.*",
    "agentlightning.*",
    "langgraph.*",
    "langchain_core.*",
]
ignore_missing_imports = true
